requirements.txt
python-dotenv==1.0.0
altair==4.2.2
streamlit==1.18.1
openai==0.27.7